refactor(agent): extract context-free winget parsers into winget_parse.go
